requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.21.0
opencv-python>=4.5.0
Pillow>=8.0.0
fastapi>=0.68.0
uvicorn[standard]>=0.15.0
python-multipart>=0.0.5
